Convert apiquery to backfill
Backfill is now an object with type and query. This allows future extensions to queries that are not backed by capi.

As a first step both `apiQuery` and `backfill` are available, the client tries to use `backfill` first and `apiQuery` if missing.

There's a change in a public method interface so I've created a new method with a different signature, users of the library have time to migrate.