Introduce ContentFormat into FAPI Client
Both Apps and Dotcom platforms currently use a `Format` type to
determine how to style content. This type is defined in the
`@guardian/types` library and, as of a recent PR, the Scala CAPI Client.

In order for this information to be used in Trails and on Cards the FAPI
client needs to be updated to use the new information from the Scala
client and pass it through to the consumers.

I here add the `Format` type to the `CuratedContent`,
`SupportingCuratedContent` and `LatestSnap` classes to allow it to be
consumed on-platform.

Changes:
- Bump CAPI client version to the most recent version as of writing
(v17.11)
- Add a `format` field to the above listed classes and create a
`ContentFormat` case class to type this field.
  - This case class is copied over from Frontend, but there is future
  work to move both `Format` and its types to the top level rather than
  in the client. At this point the duplication will no longer be
  necessary.