Update CAPI client to help evaluate content-api-scala-client PR #359
https://github.com/guardian/content-api-scala-client/pull/359 has
several medium-sized refactor-changes to its auto-pagination support,
to the extent that it will justify a new major version number, going
from v18 to v19. To increase confidence that the CAPI-client changes
aren't unexpectedly breaking things, it's good to try them out on as
many projects as possible that use the CAPI client!

We're mainly looking to see if these changes break compilation - I don't
think the facia-scala-client actually used any of the refactored
'pagination' functionality in the CAPI-client.