Add Continuous Integration (CI) as GitHub Action
Running the tests for this project requires read access to
s3://facia-tool-store/DEV/, so we need to provide the GitHub Action
with AWS credentials for a AWS role that allows that.

We're using https://github.com/aws-actions/configure-aws-credentials
to grant the credentials, and https://github.com/guardian/cdk to
create the AWS Role.

Co-authored-by: Akash Askoolum <akash.askoolum@guardian.co.uk>