Disable artifact compatibility checking for next release
As https://github.com/guardian/facia-scala-client/pull/307 introduces a
new artifact, unfortunately artifact compatibility checking will crash
out (reporting that it can't find the previous version of the brand new
artifact to check compatibility against). So, we have to disable the
check for now, and then revert this change after the new artifact is
released.