Fix test so it does not rely on JSON object field order
JSON objects have unordered keys, so there's no 'correct' order for the fields
to appear in when the JSON is converted into a String. This means it's a bad idea
to write tests that assert that JSON should match a specific string - because
the string may change with new versions of the underlying software libraries, eg
with changes to Java or Scala versions.

This particular 'should serialize territories' test in `facia-scala-client`
works at the moment, but spontaneously fails under Scala 3 with the changes
in https://github.com/guardian/facia-scala-client/pull/287 :

https://github.com/guardian/facia-scala-client/actions/runs/9894534871/job/27332773020#step:5:2147

```
[info] - should serialize territories *** FAILED ***
[info]   "...nt-content=true"},"t[argetedTerritory":"NZ","uneditable":true,"type":"news/most-popular]"}" was not equal to "...nt-content=true"},"t[ype":"news/most-popular","uneditable":true,"targetedTerritory":"NZ]"}" (ConfigSpec.scala:58)
```

A quick fix is to not make an assertion against a _string_, but to make an assertion
against the JSON _type_ (in our case, `JsValue`) which has proper equality rules, and
does not mind about _ordering_ of the fields.