Drop Scala 2.12 and Play 2.7, begin partial support for Scala 3
Fixes #317, meaning that PRs like #287 can use Scala 2.13 features.
Also dropping Play 2.7 - Play 2.8 was end-of-life (EOL) on 31st May 2024,
so dropping support for Play 2.7 seems very reasonable!

Running the tests under Scala 3 required commit 3690941, fixing a test so
that it doesn't rely on JSON object's field order.

All 7 consumers of the facia-scala-client library are now on Scala 2.13,
and Play 2.8 or above, so they do not pose blockers to dropping Scala 2.12 and Play 2.7:

* guardian/ophan - Play 3.0.4
* guardian/frontend - Play 3.0.4
* guardian/apple-news - Play 3.0.2
* guardian/editors-picks-uploader - AWS Lambda, doesn't use Play
* guardian/facia-tool - Play 3.0.2
* guardian/story-packages - Play 2.8.19
* guardian/mobile-apps-api - Play 3.0.3