Fix release so that it does not attempt to use 2.12
After https://github.com/guardian/facia-scala-client/pull/318 dropped
support for Scala 2.12, release 8.0.1 correctly no longer published Scala 2.12
artifacts, but unfortunately the 'root' project (which does not have any
code!) was still defaulting to using Scala 2.12, because that is the
default sbt uses if no `scalaVersion` is set.

The next time we tried to do a release (https://github.com/guardian/facia-scala-client/pull/319#issuecomment-2263535032)
the automated-compatibility checking failed, because it saw that one project
was using Scala 2.12, and so it attempted to download all the non-existent
Scala 2.12 artifacts for release 8.0.1 - and failed.

To fix this, we can simply ensure that all modules in this sbt build default
to Scala 2.13, not 2.12 - ie particularly, the 'root' module!